Added Autonomous toggle button and sending serial to controller

+++ mcp-vs-main/src/Robot.h
@@ -28,7 +28,7 @@
 
 void pwm_init(void);
 void motorDrive();
-void serialOutput(int8_t val);
+void serialOutput(int32_t val);
 void setupMotors();
 void setupRangeSensors();
 void motorDrive(int16_t *fc_ptr, int16_t *rc_ptr);
